requirements.txt
amqp==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_full_version < "3.11.3"
billiard==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
celery==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
celery[redis]==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
channels-redis[cryptography]==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
channels==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.10" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
croniter==2.0.5 ; python_version >= "3.10" and python_version < "4.0"
cryptography==42.0.8 ; python_version >= "3.10" and python_version < "4.0"
django-celery-results==2.5.1 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.10" and python_version < "4.0"
django-dbbackup==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
django-filter==23.5 ; python_version >= "3.10" and python_version < "4.0"
django-redis==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
django-rq==2.10.2 ; python_version >= "3.10" and python_version < "4.0"
django-stubs-ext==5.0.2 ; python_version >= "3.10" and python_version < "4.0"
django-tasks-scheduler==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
django-tasks==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-api-key==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
drf-url-filters==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
eventlet==0.35.2 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.1 ; python_version >= "3.10" and python_version < "3.11"
greenlet==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
kombu==5.3.7 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.6 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.46 ; python_version >= "3.10" and python_version < "4.0"
psutil==5.9.8 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
redis==5.0.4 ; python_version >= "3.10" and python_version < "4.0"
rq==1.16.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.27.1 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
voluptuous==0.14.2 ; python_version >= "3.10" and python_version < "4.0"
watchdog==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.10" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.6.0 ; python_version >= "3.10" and python_version < "4.0"
